rename BaseScraper to BaseScraperWithBrowser
